feat(v1-12): Part C+D — AppShell职责解耦与全窗口组合Story
- AppShell.tsx 从 1,267 行按职责解耦为 9 个文件（每个 ≤258 行）
- 提取 useAppShellLayout + useAppShellKeyboard hooks
- 提取 useAppShellAiCompare hook（AI 比较状态管理）
- 提取 appShellCommands（命令面板条目构建）
- 提取 AppShellOverlays（覆盖层对话框 + ZenMode）
- 提取 AppShellLeftPanel + AppShellMainArea + AppShellRightPanel
- AppShell.tsx 精简至 149 行纯编排层
- 更新 panel-orchestrator guard 测试适配新文件结构
- 新增 3 个 fullscreen composition stories

AC: 19-23, 15, 16, 18
